fix: incorrect config.upload types
